Ajout ex21 POO

+++ Dev Victor/Ex POO/Ex21/Program.cs
@@ -1,6 +1,6 @@
 ﻿using Ex21.Exceptions;
 
-static int calculMoyenne(int[] notes)
+static double calculMoyenne(int[] notes)
 {
     int a = 0;
     for (int i = 0; i < notes.Length; i++)
